+++ apps/backend/src/libs/enums/enums.ts
@@ -1,8 +1,8 @@
+export { SortingOrder } from "./sorting-order.enum.js";
 export {
 	APIErrorType,
 	APIPath,
 	AppEnvironment,
 	AuthApiPath,
-	AuthExceptionMessage,
 	HTTPCode,
 } from "@smartscapes/shared";