+++ apps/backend/src/libs/modules/database/libs/enums/database-table-name.enum.ts
@@ -2,9 +2,14 @@ const DatabaseTableName = {
 	GROUPS: "groups",
 	GROUPS_TO_PERMISSIONS: "groups_to_permissions",
 	MIGRATIONS: "migrations",
+	NOTIFICATIONS: "notifications",
 	PERMISSIONS: "permissions",
 	POINTS_OF_INTEREST: "points_of_interest",
+	REVIEWS: "reviews",
 	ROUTE_CATEGORIES: "route_categories",
+	ROUTES: "routes",
+	ROUTES_TO_POIS: "routes_to_pois",
+	USER_FOLLOWS: "user_follows",
 	USERS: "users",
 } as const;