+++ apps/backend/src/modules/notifications/libs/types/types.ts
@@ -1,0 +1,4 @@
+export {
+	type NotificationCreateRequestDto,
+	type NotificationGetAllItemResponseDto,
+} from "@smartscapes/shared";