+++ apps/backend/src/modules/points-of-interest/points-of-interest.service.ts
@@ -1,3 +1,4 @@
+import { METERS_IN_KM } from "~/libs/constants/constants.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
 import { type CollectionResult, type Service } from "~/libs/types/types.js";
 import { PointsOfInterestEntity } from "~/modules/points-of-interest/points-of-interest.entity.js";
@@ -6,6 +7,7 @@ import { type PointsOfInterestRepository } from "~/modules/points-of-interest/po
 import { PointOfInterestExceptionMessage } from "./libs/enums/enums.js";
 import { PointOfInterestError } from "./libs/exceptions/exceptions.js";
 import {
+	type PointsOfInterestFindAllOptions,
 	type PointsOfInterestRequestDto,
 	type PointsOfInterestResponseDto,
 } from "./libs/types/type.js";
@@ -47,15 +49,34 @@ class PointsOfInterestService implements Service {
 		return true;
 	}
 
-	public async findAll(): Promise<
-		CollectionResult<PointsOfInterestResponseDto>
-	> {
-		const items = await this.pointsOfInterestRepository.findAll();
+	public async findAll(
+		options: null | PointsOfInterestFindAllOptions,
+	): Promise<CollectionResult<PointsOfInterestResponseDto>> {
+		const hasLocationFilter =
+			options && Boolean(options.latitude) && Boolean(options.longitude);
+
+		if (!hasLocationFilter) {
+			const items = await this.pointsOfInterestRepository.findAll(options);
+
+			return {
+				items: items.map((item) => item.toObject()),
+			};
+		}
+
+		const DEFAULT_SEARCH_RADIUS_KM = 5;
+
+		const { radius = DEFAULT_SEARCH_RADIUS_KM } = options;
+
+		const searchParameters = {
+			...options,
+			radius: radius * METERS_IN_KM,
+		};
+
+		const items =
+			await this.pointsOfInterestRepository.findNearby(searchParameters);
 
 		return {
-			items: items.map((item) => {
-				return item.toObject();
-			}),
+			items: items.map((item) => item.toObject()),
 		};
 	}