+++ apps/backend/src/modules/users/libs/types/user-password-details.type.ts
@@ -1,5 +1,7 @@
 type UserPasswordDetails = {
+	firstName: string;
 	id: number;
+	lastName: string;
 	passwordHash: string;
 	passwordSalt: string;
 };