+++ apps/frontend/src/libs/components/app/app.tsx
@@ -1,27 +1,19 @@
-import {
-	Button,
-	Header,
-	Loader,
-	RouterOutlet,
-	Sidebar,
-} from "~/libs/components/components.js";
-import { NAVIGATION_ITEMS } from "~/libs/constants/constants.js";
+import { Loader, RouterOutlet } from "~/libs/components/components.js";
 import { AppRoute } from "~/libs/enums/enums.js";
-import { useAppDispatch, useEffect, useLocation } from "~/libs/hooks/hooks.js";
+import {
+	useAppDispatch,
+	useAppSelector,
+	useEffect,
+	useLocation,
+} from "~/libs/hooks/hooks.js";
+import { actions as authActions } from "~/modules/auth/auth.js";
 import { actions as userActions } from "~/modules/users/users.js";
 
-import { mockImages } from "../carousel/assets/mock-images/mock-images.js";
-import { Carousel } from "../carousel/carousel.js";
 import styles from "./styles.module.css";
 
 const App = (): React.JSX.Element => {
 	const { pathname } = useLocation();
 	const dispatch = useAppDispatch();
-	const mockUserWithoutAvatar = {
-		avatarUrl: null,
-		firstName: "John",
-		lastName: "Smith",
-	};
 
 	const isRoot = pathname === AppRoute.APP;
 
@@ -31,26 +23,25 @@ const App = (): React.JSX.Element => {
 		}
 	}, [isRoot, dispatch]);
 
+	useEffect(() => {
+		void dispatch(authActions.getAuthenticatedUser());
+	}, [dispatch]);
+
+	const isInitialized = useAppSelector(({ auth }) => auth.isInitialized);
+
+	if (!isInitialized) {
+		return (
+			<div className={styles["loader-container"]}>
+				<Loader />
+			</div>
+		);
+	}
+
 	return (
 		<div className={styles["container"]}>
 			<div className={styles["outlet-container"]}>
 				<RouterOutlet />
 			</div>
-			{isRoot && (
-				<div className={styles["components-container"]}>
-					<Header user={mockUserWithoutAvatar} />
-					<div className={styles["sidebar-container"]}>
-						<Sidebar navigationItems={NAVIGATION_ITEMS} />
-					</div>
-					<Loader />
-					<div className={styles["button-container"]}>
-						<Button label="Button for test" type="button" />
-					</div>
-					<div className={styles["carousel-container"]}>
-						<Carousel images={mockImages} />
-					</div>
-				</div>
-			)}
 		</div>
 	);
 };