+++ apps/frontend/src/libs/components/carousel/libs/helpers/helpers.ts
@@ -1,2 +1,0 @@
-export { getCarouselParameters } from "./get-carousel-parameters.helper.js";
-export { snapToNearestItem } from "./snap-to-nearest-item.helper.js";