+++ apps/frontend/src/libs/components/header/header.tsx
@@ -2,17 +2,34 @@ import { type JSX } from "react";
 
 import appLogo from "~/assets/images/logo.svg";
 import { AppRoute } from "~/libs/enums/enums.js";
+import { type UserAuthResponseDto, type ValueOf } from "~/libs/types/types.js";
 
-import { Avatar, Button, Link } from "../components.js";
-import { type User } from "./libs/types/types.js";
+import { Button, Link } from "../components.js";
+import { AuthenticatedHeader } from "./libs/components/authenticated-header/authenticated-header.js";
 import styles from "./styles.module.css";
 
 type Properties = {
-	user: null | User;
+	actions: {
+		label: string;
+		to: ValueOf<typeof AppRoute>;
+	}[];
+	user: null | UserAuthResponseDto;
 };
 
-const Header = ({ user }: Properties): JSX.Element => {
-	const hasUser = Boolean(user);
+const Header = ({ actions, user }: Properties): JSX.Element => {
+	const renderHeaderContent = (): JSX.Element => {
+		if (user) {
+			return <AuthenticatedHeader user={user} />;
+		}
+
+		return (
+			<div className={styles["buttons"]}>
+				{actions.map(({ label, to }) => (
+					<Button key={to} label={label} to={to} type="button" />
+				))}
+			</div>
+		);
+	};
 
 	return (
 		<header className={styles["header"]}>
@@ -25,19 +42,7 @@ const Header = ({ user }: Properties): JSX.Element => {
 					width={136}
 				/>
 			</Link>
-			{hasUser ? (
-				<div className={styles["user-info"]}>
-					{user && <Avatar user={user} />}
-					<div className={styles["name"]}>
-						{user?.firstName} {user?.lastName}
-					</div>
-				</div>
-			) : (
-				<div className={styles["buttons"]}>
-					<Button label="Sign up" to={AppRoute.SIGN_UP} type="button" />
-					<Button label="Sign in" to={AppRoute.SIGN_IN} type="button" />
-				</div>
-			)}
+			{renderHeaderContent()}
 		</header>
 	);
 };