+++ apps/frontend/src/modules/auth/slices/auth.slice.ts
@@ -4,16 +4,18 @@ import { type UserAuthResponseDto } from "@smartscapes/shared";
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
-import { getAuthenticatedUser, signIn, signUp } from "./actions.js";
+import { getAuthenticatedUser, logout, signIn, signUp } from "./actions.js";
 
 type State = {
-	authenticatedUser: null | Pick<UserAuthResponseDto, "email" | "id">;
+	authenticatedUser: null | UserAuthResponseDto;
 	dataStatus: ValueOf<typeof DataStatus>;
+	isInitialized: boolean;
 };
 
 const initialState: State = {
 	authenticatedUser: null,
 	dataStatus: DataStatus.IDLE,
+	isInitialized: false,
 };
 
 const { actions, name, reducer } = createSlice({
@@ -21,6 +23,7 @@ const { actions, name, reducer } = createSlice({
 		builder.addCase(getAuthenticatedUser.fulfilled, (state, action) => {
 			state.authenticatedUser = action.payload ? action.payload.data : null;
 			state.dataStatus = DataStatus.FULFILLED;
+			state.isInitialized = true;
 		});
 		builder.addCase(getAuthenticatedUser.pending, (state) => {
 			state.dataStatus = DataStatus.PENDING;
@@ -28,6 +31,12 @@ const { actions, name, reducer } = createSlice({
 		builder.addCase(getAuthenticatedUser.rejected, (state) => {
 			state.authenticatedUser = null;
 			state.dataStatus = DataStatus.REJECTED;
+			state.isInitialized = true;
+		});
+
+		builder.addCase(logout.fulfilled, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.FULFILLED;
 		});
 
 		builder.addCase(signUp.fulfilled, (state, action) => {