+++ CHANGELOG.md
@@ -1,5 +1,121 @@
 # Changelog
 
+## [1.19.1](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.19.0...v1.19.1) (2025-08-08)
+
+
+### Bug Fixes
+
+* **root:** fix database erDiagram ss-272 ([#273](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/273)) ([b4d379e](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/b4d379e3fa7daef3ba9b73f8d6a7ea50e3fa7eaa))
+
+## [1.19.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.18.1...v1.19.0) (2025-08-07)
+
+
+### Features
+
+* **frontend:** add floating sidebar ss-106 ([#214](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/214)) ([d576265](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/d57626569c8a05eb1b35fe56b90760304b4bef12))
+
+## [1.18.1](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.18.0...v1.18.1) (2025-08-07)
+
+
+### Bug Fixes
+
+* **frontend:** app component waits for get-authicated-user dispatch ss-247 ([#252](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/252)) ([30cc3dc](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/30cc3dc068a407ca02389ed1724f0b853732ec71))
+
+## [1.18.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.17.2...v1.18.0) (2025-08-07)
+
+
+### Features
+
+* **frontend:** add styles to select ss-255 ([#257](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/257)) ([3a8be15](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/3a8be15b012417608f56dcc3024ddbc01fc92547))
+
+## [1.17.2](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.17.1...v1.17.2) (2025-08-07)
+
+
+### Bug Fixes
+
+* **backend:** fix poi location schema in swagger documentation ss-258 ([#259](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/259)) ([323d2cf](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/323d2cfb51ba8c817bff46f00b9d079efffacca7))
+
+## [1.17.1](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.17.0...v1.17.1) (2025-08-07)
+
+
+### Bug Fixes
+
+* **backend:** update reviews module swagger docs ss-251 ([#253](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/253)) ([4753c0e](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/4753c0e243a7d4d8b39ae271c911b0bfe5995d51))
+
+## [1.17.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.16.0...v1.17.0) (2025-08-07)
+
+
+### Features
+
+* **backend/shared:** add reviews module ss-75 ([#147](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/147)) ([435363e](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/435363e7ee382997fab923744e484ecc0efde238))
+
+## [1.16.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.15.0...v1.16.0) (2025-08-07)
+
+
+### Features
+
+* **backend/shared:** add routes module ss-69 ([#104](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/104)) ([b419013](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/b419013fd741b365045a9e6f3d49c30a1885a6b7))
+* **frontend:** update buttons in header ss-186 ([#210](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/210)) ([e4f8d7f](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/e4f8d7ffe91ea72f18bc8932a086e4c7e1194f42))
+
+## [1.15.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.14.0...v1.15.0) (2025-08-06)
+
+
+### Features
+
+* **frontend:** add custom select component ss-192 ([#217](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/217)) ([8e1e395](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/8e1e3950088ad6ea842ed26ff91e5a01731b06ac))
+
+## [1.14.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.13.0...v1.14.0) (2025-08-06)
+
+
+### Features
+
+* **backend:** add security to swagger documentation of notifications ss-233 ([#234](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/234)) ([923e93e](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/923e93e1b2194455590794f45fe1ce9ff87bf598))
+* **frontend:** frontend protected routes ss-24 ([#199](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/199)) ([ad96806](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/ad96806486f6168a4f49d9807544107c99e5f37e))
+
+## [1.13.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.12.0...v1.13.0) (2025-08-06)
+
+
+### Features
+
+* **backend:** add super admin seed ss-187 ([#212](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/212)) ([15457e0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/15457e0521ea078d179bd0471402713f92b9b751))
+
+## [1.12.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.11.2...v1.12.0) (2025-08-06)
+
+
+### Features
+
+* **backend:** search closest points to authenticated user ss-188 ([#198](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/198)) ([144e41c](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/144e41c87e7f4fbf313cac3fc2ac6c53464949e9))
+* **frontend:** toast notification smoother show ss-181 ([#191](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/191)) ([3328e0c](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/3328e0c29f1ef63efb20c08dfd04982439c0c924))
+
+
+### Bug Fixes
+
+* **backend:** introduce swagger doc for dto in notifications ss-229 ([#230](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/230)) ([20f10d8](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/20f10d8877c683e15920ca354b5accfe114c9f9f))
+
+## [1.11.2](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.11.1...v1.11.2) (2025-08-06)
+
+
+### Bug Fixes
+
+* **backend:** correct value of type ss-226 ([#227](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/227)) ([085b1b9](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/085b1b908a9575d4b0394c7fb6f8b3cfc414e926))
+
+## [1.11.1](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.11.0...v1.11.1) (2025-08-05)
+
+
+### Bug Fixes
+
+* **backend:** update swagger sign-in response example ss-221 ([#222](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/222)) ([7fc47cc](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/7fc47ccaeef644774be314cb7b2fa5be3b04d1d2))
+
+## [1.11.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.10.0...v1.11.0) (2025-08-05)
+
+
+### Features
+
+* **frontend:** add custom scrollbar styling ss-180 ([#190](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/190)) ([c5558dc](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/c5558dcb27c3fe0752035b8839c5addc94a9421e))
+* **frontend:** carousel improvement ss-183 ([#194](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/194)) ([da750df](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/da750df01137a81ef0298728535ca9c92ad5c202))
+* **frontend:** enable PWA for frontend ss-68 ([#89](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/89)) ([979f3f4](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/979f3f4d8eaad080b64ff2532ee0f18a099e5953))
+* **shared/backend:** add notifications module ss-73 ([#131](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/131)) ([e2b4a15](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/e2b4a1514dd0a10f89a36d8bc3894c61444f168f))
+
 ## [1.10.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/v1.9.3...v1.10.0) (2025-08-04)