+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,63 @@
 # Changelog
 
+## [1.16.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.15.1...frontend-v1.16.0) (2025-08-07)
+
+
+### Features
+
+* **frontend:** add floating sidebar ss-106 ([#214](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/214)) ([d576265](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/d57626569c8a05eb1b35fe56b90760304b4bef12))
+
+## [1.15.1](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.15.0...frontend-v1.15.1) (2025-08-07)
+
+
+### Bug Fixes
+
+* **frontend:** app component waits for get-authicated-user dispatch ss-247 ([#252](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/252)) ([30cc3dc](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/30cc3dc068a407ca02389ed1724f0b853732ec71))
+
+## [1.15.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.14.0...frontend-v1.15.0) (2025-08-07)
+
+
+### Features
+
+* **frontend:** add styles to select ss-255 ([#257](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/257)) ([3a8be15](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/3a8be15b012417608f56dcc3024ddbc01fc92547))
+
+## [1.14.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.13.0...frontend-v1.14.0) (2025-08-07)
+
+
+### Features
+
+* **frontend:** update buttons in header ss-186 ([#210](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/210)) ([e4f8d7f](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/e4f8d7ffe91ea72f18bc8932a086e4c7e1194f42))
+
+## [1.13.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.12.0...frontend-v1.13.0) (2025-08-06)
+
+
+### Features
+
+* **frontend:** add custom select component ss-192 ([#217](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/217)) ([8e1e395](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/8e1e3950088ad6ea842ed26ff91e5a01731b06ac))
+
+## [1.12.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.11.0...frontend-v1.12.0) (2025-08-06)
+
+
+### Features
+
+* **frontend:** frontend protected routes ss-24 ([#199](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/199)) ([ad96806](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/ad96806486f6168a4f49d9807544107c99e5f37e))
+
+## [1.11.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.10.0...frontend-v1.11.0) (2025-08-06)
+
+
+### Features
+
+* **frontend:** toast notification smoother show ss-181 ([#191](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/191)) ([3328e0c](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/3328e0c29f1ef63efb20c08dfd04982439c0c924))
+
+## [1.10.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.9.3...frontend-v1.10.0) (2025-08-05)
+
+
+### Features
+
+* **frontend:** add custom scrollbar styling ss-180 ([#190](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/190)) ([c5558dc](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/c5558dcb27c3fe0752035b8839c5addc94a9421e))
+* **frontend:** carousel improvement ss-183 ([#194](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/194)) ([da750df](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/da750df01137a81ef0298728535ca9c92ad5c202))
+* **frontend:** enable PWA for frontend ss-68 ([#89](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/89)) ([979f3f4](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/979f3f4d8eaad080b64ff2532ee0f18a099e5953))
+
 ## [1.9.3](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/frontend-v1.9.2...frontend-v1.9.3) (2025-08-01)